src/core/nova_engine.py
"""
Core Nova Act Engine module.

This module provides the foundation for browser automation using Amazon Nova Act.
"""

import asyncio
import logging
from datetime import datetime
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

logger = logging.getLogger(__name__)

class NovaActEngine:
    """
    Core Nova Act Engine for browser automation.
    """
    
    def __init__(self, config):
        """
        Initialize the Nova Act Engine.
        
        Args:
            config (dict): System configuration
        """
        self.config = config
        self.api_key = os.getenv("NOVA_ACT_API_KEY")
        self.browser = None
        self.session_start_time = None
        
        if not self.api_key:
            logger.error("Nova Act API key not found in environment variables")
            raise ValueError("Nova Act API key is required")
        
        # Placeholder for the actual Nova Act import
        # from nova_act import NovaAct
        # self.nova = NovaAct(api_key=self.api_key)
        
        logger.info("Nova Act Engine initialized")
    
    async def initialize_browser(self):
        """
        Initialize a new browser session.
        
        Returns:
            object: Browser instance
        """
        logger.info("Initializing browser session")
        
        # Placeholder for actual implementation
        # self.browser = await self.nova.start_browser(
        #     headless=self.config["nova_act"]["headless"]
        # )
        
        # For now, just create a mock browser
        self.browser = {"initialized": True}
        self.session_start_time = datetime.now()
        
        logger.info("Browser session initialized")
        return self.browser
    
    async def is_session_valid(self):
        """
        Check if the current session is still valid.
        
        Returns:
            bool: True if session is valid, False otherwise
        """
        if not self.browser or not self.session_start_time:
            return False
        
        session_age = (datetime.now() - self.session_start_time).total_seconds() / 60
        return session_age < self.config["processing"]["session_timeout"]
    
    async def ensure_browser(self):
        """
        Ensure a valid browser session exists.
        
        Returns:
            object: Browser instance
        """
        if not await self.is_session_valid():
            if self.browser:
                await self.close_browser()
            await self.initialize_browser()
        return self.browser
    
    async def close_browser(self):
        """
        Close the browser session.
        """
        if self.browser:
            # Placeholder for actual implementation
            # await self.browser.close()
            
            logger.info("Browser session closed")
            self.browser = None
            self.session_start_time = None
    
    async def execute_action(self, action_fn, max_retries=None):
        """
        Execute an action with retry logic.
        
        Args:
            action_fn (callable): Async function to execute
            max_retries (int, optional): Maximum number of retry attempts
            
        Returns:
            Any: Result of the action
            
        Raises:
            Exception: If all retry attempts fail
        """
        if max_retries is None:
            max_retries = self.config["nova_act"]["retry_attempts"]
            
        attempts = 0
        last_error = None
        
        while attempts < max_retries:
            try:
                browser = await self.ensure_browser()
                result = await action_fn(browser)
                return result
            except Exception as e:
                last_error = e
                attempts += 1
                logger.warning(f"Action failed (attempt {attempts}/{max_retries}): {str(e)}")
                
                if attempts >= max_retries:
                    break
                
                # Implement exponential backoff
                backoff_time = 2 ** attempts
                logger.info(f"Retrying in {backoff_time} seconds...")
                await asyncio.sleep(backoff_time)
        
        # If we reach here, all attempts failed
        logger.error(f"Action failed after {max_retries} attempts: {str(last_error)}")
        raise Exception(f"Action failed after {max_retries} attempts: {last_error}")
